+++ package.json
@@ -55,6 +55,7 @@
 		"@types/jest": "^29.5.12",
 		"@types/json-schema": "^7.0.15",
 		"@types/node": "^20.11.0",
+		"dotenv": "^16.4.7",
 		"jest": "^29.7.0",
 		"shx": "^0.3.4",
 		"ts-jest": "^29.1.2",
